add separate folder for book detail related components
