+++ assets/src/libraries/Book.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import Paper from '@material-ui/core/Paper';
 import Button from '@material-ui/core/Button';
 import parse from 'url-parse';
-import BookDetail from './BookDetail';
+import BookDetail from './detail/BookDetail';
 import { borrowCopy, returnBook, joinWaitlist } from '../services/BookService';
 import BookModel from '../models/Book';
