+++ assets/src/libraries/detail/BookBorrowers.jsx
